+++ examples/client/apis.js
@@ -14,7 +14,8 @@ export const getInfo = get('api/getInfo', null, {
 
 export const getInfoCustom = get('/api/getInfoCustom', {
   msgKey: 'error_msg',
-  codeKey: 'dm_error'
+  codeKey: 'dm_error',
+  successCode: 0
 })
 
 export const postInfo = post('api/postInfo')
@@ -35,4 +36,13 @@ export const getGitHubUser = gitHubRestFulGet('users/$user', {
 export const getInfoWithMock = mockGetInfo(get('api/getInfo', null, {
   autoLoading: false
 }))
+
+class Apis {
+  @mockGetInfo
+  getInfoWithMock = get('api/getInfo', null, {
+    autoLoading: false
+  })
+}
+
+export const apis = new Apis()